move : 컴포넌트로 폴더 이동
